docs: divided the code into ES modules
