+++ js/random-num.js
@@ -1,0 +1,4 @@
+function randomNum(min, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+export {randomNum};
